editProfileData (versao mais atualizada)
(widths corretas e funçoes na profilePage)
